src/voice/audio_player.py
"""Audio playback system"""
import logging

class AudioPlayer:
    """Audio playback via pygame"""
    def __init__(self):
        self.logger = logging.getLogger("AudioPlayer")
        try:
            import pygame
            pygame.mixer.init()
            self.pygame = pygame
            self.logger.info("✅ Audio player initialized")
        except ImportError:
            self.logger.warning("⚠️  pygame not installed")
            self.pygame = None

    async def play(self, audio_bytes: bytes):
        """Play audio bytes"""
        if not audio_bytes or not self.pygame:
            return
        # Playback implementation
        self.logger.info("🔊 Playing audio")
